decouple Gyro Space codeset to a header
Moved all Gyro Space Transformer code to a dedicated header for easier organization and porting code set.
